requirements.txt
aiogram
SQLAlchemy
google-generativeai
python-docx
python-dotenv
psycopg[binary]
aiofiles